Modify: mock data 생성 및 리팩토링 작업
