+++ public/data/boyoonKim/commentData.json
@@ -1,0 +1,17 @@
+[
+  {
+    "id": 1,
+    "name": "wecode",
+    "contents": "Welcome to world best coding bootcamp!"
+  },
+  {
+    "id": 2,
+    "name": "boyoon",
+    "contents": "나만의 속도가 있다!"
+  },
+  {
+    "id": 3,
+    "name": "jungkook",
+    "contents": "보윤님 화이팅!"
+  }
+]
