+++ src/pages/boyoonKim/Main/CommentsBox.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faHeart } from '@fortawesome/free-regular-svg-icons';
 import './../Main/MainBo.scss';
@@ -7,8 +7,9 @@ import './../../../styles/reset.scss';
 
 const CommentsBox = ({ value, index }) => {
   // props 자리에서 바로 비구조화 할당
+
   return (
-    <li key={index}>
+    <li className="commentLi" key={index}>
       <div>
         <span className="userId">아이디</span>
         <span clsssName="commentsContents">{value}</span>
